move getAllTokens from esdt.service to token.service
