+++ src/endpoints/network/network.service.ts
@@ -18,8 +18,8 @@ import { CacheInfo } from 'src/utils/cache.info';
 import { GatewayComponentRequest } from 'src/common/gateway/entities/gateway.component.request';
 import { NumberUtils, CachingService, ApiService } from '@elrondnetwork/erdnest';
 import { About } from './entities/about';
-import { EsdtService } from '../esdt/esdt.service';
 import { PluginService } from 'src/common/plugins/plugin.service';
+import { TokenService } from '../tokens/token.service';
 
 @Injectable()
 export class NetworkService {
@@ -38,9 +38,8 @@ export class NetworkService {
     private readonly apiService: ApiService,
     @Inject(forwardRef(() => StakeService))
     private readonly stakeService: StakeService,
-    @Inject(forwardRef(() => EsdtService))
-    private readonly esdtService: EsdtService,
     private readonly pluginService: PluginService,
+    private readonly tokenService: TokenService,
   ) { }
 
   async getConstants(): Promise<NetworkConstants> {
@@ -149,7 +148,7 @@ export class NetworkService {
         'getTotalStakeByType',
       ),
       this.dataApiService.getQuotesHistoricalLatest(DataQuoteType.price),
-      this.esdtService.getTokenMarketCapRaw(),
+      this.tokenService.getTokenMarketCapRaw(),
     ]);
